Atelier 6 - Vues - grille, case, entetes Fait:  - La grille se crée avec les parametre recut  - les fichier son construit avec les méthode et leur signatures TODO:  - Ajouter les jouer a la vue partie (barre de droite)  - prendre les parametre pour la construiction de la grille correctement  - ajouter les listener a la grille

+++ app/src/main/java/ca/cours5b5/justinfofana/global/GConstantes.java
@@ -23,10 +23,10 @@ public class GConstantes {
     public static final int GRILLE_MARGIN = 5;
 
     public static final int RANGE_ENTETE = 0;
-    public static final int ENTETE_POIDS_COLONNE = 1;
-    public static final int ENTETE_POIDS_RANGE = 3;
-    public static final int CASE_POIDS_COLONNE = 1;
-    public static final int CASE_POIDS_RANGE = 1;
+    public static final float ENTETE_POIDS_COLONNE = 1;
+    public static final float ENTETE_POIDS_RANGE = 3;
+    public static final float CASE_POIDS_COLONNE = 1;
+    public static final float CASE_POIDS_RANGE = 1;
 
 
 }
